API Gateway fully working and refactored price variable - it should be a double

+++ src/main/java/com/siramiks/PaymentService/entity/Payment.java
@@ -37,7 +37,7 @@ public class Payment {
   private LocalDateTime createdAt;
 
   @Column(name = "payment_total")
-  private long totalPayment;
+  private double totalPayment;
 
   @Column(name = "payment_method")
   private String paymentMethod;
